src/contoso-travel-maf/requirements.txt
azure-ai-agentserver-agentframework==1.0.0b16
azure-ai-projects>=2.0.0
azure-identity
python-dotenv
pandas
